+++ android-example/src/main/java/org/rewedigital/katana/android/example/fragment/FirstFragmentModule.kt
@@ -1,12 +1,11 @@
 package org.rewedigital.katana.android.example.fragment
 
-import org.rewedigital.katana.bind
 import org.rewedigital.katana.createModule
-import org.rewedigital.katana.factory
+import org.rewedigital.katana.dsl.compact.factory
 
 const val FRAGMENT_DEPENDENCY1 = "FRAGMENT_DEPENDENCY1"
 
 val firstFragmentModule = createModule {
 
-    bind<String>(FRAGMENT_DEPENDENCY1) { factory { "FRAGMENT_DEPENDENCY1" } }
+    factory(name = FRAGMENT_DEPENDENCY1) { "FRAGMENT_DEPENDENCY1" }
 }