+++ android-example/src/main/java/org/rewedigital/katana/android/example/fragment/SecondFragmentModule.kt
@@ -1,16 +1,15 @@
 package org.rewedigital.katana.android.example.fragment
 
 import org.rewedigital.katana.android.example.fragment.inject.Container
-import org.rewedigital.katana.bind
 import org.rewedigital.katana.createModule
-import org.rewedigital.katana.factory
-import org.rewedigital.katana.get
+import org.rewedigital.katana.dsl.compact.factory
+import org.rewedigital.katana.dsl.get
 
 const val FRAGMENT_DEPENDENCY2 = "FRAGMENT_DEPENDENCY2"
 
 val secondFragmentModule = createModule {
 
-    bind<String>(FRAGMENT_DEPENDENCY2) { factory { "FRAGMENT_DEPENDENCY2" } }
+    factory(name = FRAGMENT_DEPENDENCY2) { "FRAGMENT_DEPENDENCY2" }
 
-    bind<Container> { factory { Container(get(SOME_DEPENDENCY), get(FRAGMENT_DEPENDENCY2)) } }
+    factory { Container(get(SOME_DEPENDENCY), get(FRAGMENT_DEPENDENCY2)) }
 }