+++ android-example/src/main/java/org/rewedigital/katana/android/example/inject/AndroidModule.kt
@@ -3,10 +3,9 @@ package org.rewedigital.katana.android.example.inject
 import android.content.Context
 import android.content.res.Resources
 import org.rewedigital.katana.android.modules.APPLICATION_CONTEXT
-import org.rewedigital.katana.bind
 import org.rewedigital.katana.createModule
-import org.rewedigital.katana.factory
-import org.rewedigital.katana.get
+import org.rewedigital.katana.dsl.compact.factory
+import org.rewedigital.katana.dsl.get
 
 /**
  * This module may provide Android specific classes like [Resources],
@@ -16,5 +15,5 @@ import org.rewedigital.katana.get
  */
 val androidModule = createModule {
 
-    bind<Resources> { factory { get<Context>(APPLICATION_CONTEXT).resources } }
+    factory { get<Context>(APPLICATION_CONTEXT).resources }
 }