+++ core/src/main/kotlin/org/rewedigital/katana/dsl/ModuleDslMarker.kt
@@ -1,0 +1,4 @@
+package org.rewedigital.katana.dsl
+
+@DslMarker
+annotation class ModuleDslMarker