+++ core/src/test/kotlin/org/rewedigital/katana/OverrideTests.kt
@@ -2,6 +2,9 @@ package org.rewedigital.katana
 
 import org.amshove.kluent.shouldNotThrow
 import org.amshove.kluent.shouldThrow
+import org.rewedigital.katana.dsl.classic.bind
+import org.rewedigital.katana.dsl.classic.factory
+import org.rewedigital.katana.dsl.get
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe