+++ core/src/test/kotlin/org/rewedigital/katana/TypeErasureTests.kt
@@ -2,6 +2,8 @@ package org.rewedigital.katana
 
 import org.amshove.kluent.shouldEqual
 import org.amshove.kluent.shouldThrow
+import org.rewedigital.katana.dsl.classic.bind
+import org.rewedigital.katana.dsl.classic.factory
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe